refactor(setup): de-part clusters, split near-cap libs, no_part gate (Refs #4086)
Wave 5: convert remaining part libraries to explicit imports, topic-split
naming/orchestrator and advanced-start test support, and ratchet CI with
repo.setup_no_part_directives.
