+++ SPEC/program/repo-lint.md
@@ -102,8 +102,9 @@
 | `tool/check_setup_dedup_seaboard_nearest_tile.dart` | Forbid a second nearest-seaboard nested full-grid + `tileAdjacentToSeaZone` Manhattan scan (`_nearestCoastalTileInProvinceForSeaZone`, `_closestSeaboardTileInProvince`, or equivalent) outside `setup_road_wiring.dart`; call sites must use `nearestSeaboardTileInProvinceForSeaZone` (Refs #4020); rule `repo.setup_dedup_seaboard_nearest_tile` |
 | `tool/check_setup_dedup_full_grid_tile_scans.dart` | Forbid nested full-grid `for (var y = 0; y < *.height` + `for (var x = 0; x < *.width` walks under `packages/colonizethis_setup/lib/src/setup/**` outside `tile_cell_scan.dart`; production full-grid consumers must use `forEachTileCell` / `forEachProvinceCell` / `visitGpOwLandTiles` (edge scans without a nested width loop remain allowed); comment lines are ignored (Refs #4029); rule `repo.setup_dedup_full_grid_tile_scans` |
 | `tool/check_setup_dedup_advanced_start_flood_fill.dart` | Forbid a second hand-rolled province `visited`/`queue`/`head` BFS (or `_enqueueUnvisitedNeighbors`) in `advanced_start_bootstrap_*.dart` outside `advanced_start_nw_topology.dart`; colonization and other bootstrap callers must use `advancedStartFloodFillProvinces` (Refs #4029); rule `repo.setup_dedup_advanced_start_flood_fill` |
-| `tool/check_setup_dedup_ownership_paint.dart` | Forbid re-inlined ownership paint scaffolding under `game_setup_ownership*.dart` outside `game_setup_ownership_paint.dart`: direct `assignTerritoriesLockedOnLandmass` / `assignTerritoriesByBfsGrowth` wiring or `_lockedGrowthOrder` clones; call sites must use `_paintLandmass` (`locked`/`bfs`); locked assigner and BFS growth engines remain separate allowlisted backends (Refs #4029); rule `repo.setup_dedup_ownership_paint` |
+| `tool/check_setup_dedup_ownership_paint.dart` | Forbid re-inlined ownership paint scaffolding under `game_setup_ownership*.dart` outside `game_setup_ownership_paint.dart`: direct `assignTerritoriesLockedOnLandmass` / `assignTerritoriesByBfsGrowth` wiring or `_lockedGrowthOrder` clones; call sites must use `paintLandmass` / `_paintLandmass` (`locked`/`bfs`); locked assigner and BFS growth engines remain separate allowlisted backends (Refs #4029, #4086); rule `repo.setup_dedup_ownership_paint` |
 | `tool/check_setup_helpers_first_class_libraries.dart` | Require `game_setup_helpers_{towns,naming,bootstrap}.dart` to remain first-class libraries (no `part of`) and `game_setup_helpers.dart` to re-export them without `part` directives (Refs #4029); rule `repo.setup_helpers_first_class_libraries` |
+| `tool/check_setup_no_part_directives.dart` | Forbid Dart `part` / `part of` directives under `packages/colonizethis_setup/lib/**` so ownership / create / capital / redistribution / locked-assigner (and naming / init orchestrator topic libs) stay proper libraries with explicit imports (mirrors orders/turn/diplomacy no-part gates; Refs #4086) — see `SPEC/program/setup-no-part-directives.md`; rule `repo.setup_no_part_directives` |
 | `tool/check_setup_test_use_shared_fixtures.dart` | Forbid re-inlined empty `Game(` / `WorldState(` shells and full `GameSetupConfig(` rebuilds under `packages/colonizethis_setup/test/**`; empty/minimal shells must use `package:colonizethis_test/game_test_fixtures.dart` `TestFixtures` (or a thin support delegate), and config variants must use `configWithOverrides` / `lockedFullInitConfig` in `init_game_orchestrator_test_support.dart` (that support file alone may construct `GameSetupConfig`); full-line comments ignored (Refs #4029). Additionally forbid running `runInitGame(config: X, options: defaultInitOptions)` with an identical config expression in two or more distinct test bodies of one file — such assert-only repeats must reuse the memoized `sharedInitGameResult(config)` harness; determinism double runs inside a single test body, calls with custom `InitGameOptions` or extra arguments (for example `generateRegion:`), and the support file itself are structurally exempt (Refs #4054); rule `repo.setup_test_use_shared_fixtures` |
 | `tool/check_orders_file_size.dart` | Enforce the two orders hot files `packages/colonizethis_orders/lib/src/orders/orders_application.dart` and `.../orders_application_completed_work.dart` at **1000 non-comment lines or fewer** each (fail when strictly greater), reusing `countNonCommentLinesFromSource`; the gate also fails when either named file is missing so it cannot silently rot if the orders module moves. Complements the repository-wide `repo.dart_file_non_comment_line_size` cap with an orders-specific message so the #3404 dedup/abstraction refactors keep these files bounded (Refs #3404); rule `repo.orders_file_size` |
 | `tool/check_orders_test_no_duplicate_fixtures.dart` | Forbid re-introducing a local copy of the shared game/world test fixtures anywhere under `packages/colonizethis_orders/test/**`: any file named `test_fixtures.dart` or any test file that redeclares a top-level `TestFixtures` class is a violation. Orders tests must import the canonical `package:colonizethis_test/game_test_fixtures.dart` factories (single source of truth) rather than duplicating them; references to the shared `TestFixtures` symbol are unaffected. Complements the testing-rule shared-fixtures policy (Refs #3714); rule `repo.orders_test_no_duplicate_fixtures` |
@@ -507,6 +508,8 @@ Do **not** add new top-level `tool/check_*.dart` **entrypoints** for CI without 
 - Given a `packages/colonizethis_ai/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4079).
 - Given the `colonizethis_ai_contracts` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4084).
 - Given a `packages/colonizethis_ai_contracts/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4084).
+- Given the `colonizethis_setup` package on `dev` where no `lib/**` file uses a `part` / `part of` directive, when `dart run tool/ct_repo_lint.dart` runs rule `repo.setup_no_part_directives`, then the rule passes without violations and exits `0` (Refs #4086).
+- Given a `packages/colonizethis_setup/lib/**` Dart file that contains `part 'foo.dart';` or `part of 'foo.dart';`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.setup_no_part_directives`, then the run fails listing each offending `path:line` and exits `1` (Refs #4086).
 - Given the H8 seller acquisition-target and OW prospect localization / mineral-eligibility pins import `flaggedSellerGame` / `supplierGame` from `packages/colonizethis_ai_contracts/test/support/` and do not redeclare `_flaggedSellerGame` / `_supplierGame`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_h8_test_shared_fixtures`, then the rule passes without violations and exits `0` (Refs #4084).
 - Given a gated H8 seller acquisition-target pin that redeclares `Game _flaggedSellerGame`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_h8_test_shared_fixtures`, then the run fails listing the offending file and exits `1` (Refs #4084).
 - Given a gated OW prospect localization or mineral-eligibility pin that redeclares `Game _supplierGame`, when `dart run tool/ct_repo_lint.dart` runs rule `repo.ai_contracts_h8_test_shared_fixtures`, then the run fails listing the offending file and exits `1` (Refs #4084).
