+++ packages/colonizethis_setup/lib/src/setup/init_game_orchestrator_types.dart
@@ -1,0 +1,66 @@
+// init_game orchestration types. SPEC/program/init-game-tool.md
+// (Refs #4086 Slice C).
+
+import 'dart:typed_data';
+
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_map/colonizethis_map.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+/// Result of running init game.
+class InitGameResult {
+  const InitGameResult({
+    required this.game,
+    required this.mapPngBytes,
+    required this.markdown,
+    required this.mapViewData,
+    required this.tileMapByRegion,
+    required this.topologyByRegion,
+    required this.combinedTopology,
+    this.warpLinks = const [],
+    this.greatPowerColorOverride,
+  });
+
+  final Game game;
+  final Uint8List mapPngBytes;
+  final String markdown;
+  final InitGameMapViewData mapViewData;
+
+  /// Tile maps per region (e.g. 'oldWorld', 'newWorld'); needed for extraction.
+  final Map<String, TileMapResult> tileMapByRegion;
+
+  /// Topology per region; used by visualizers and debug tooling.
+  final Map<String, MapTopology> topologyByRegion;
+
+  /// Single topology with prefixed node ids and warp edges for resolveTurnForGame.
+  final MapTopology combinedTopology;
+
+  /// Warp zone links between OW and NW. SPEC/game/map-topology.md.
+  final List<WarpLink> warpLinks;
+
+  /// GP id → (r, g, b) used for map view; ctdev uses this when rebuilding view data.
+  final Map<String, (int r, int g, int b)>? greatPowerColorOverride;
+}
+
+/// Options for runInitGame.
+class InitGameOptions {
+  const InitGameOptions({
+    this.cellSize = 24,
+    this.skipFillLakes = false,
+    this.renderPng = true,
+    this.greatPowerColorOverride,
+  });
+
+  final int cellSize;
+
+  /// When true, forward skipFillLakes to TileMapParams so Pass 4 (fill lakes)
+  /// is skipped in tile-map generation for both Old World and New World.
+  final bool skipFillLakes;
+
+  /// When false, skips PNG rendering inside runInitGame; mapViewData and
+  /// markdown are still produced.
+  final bool renderPng;
+
+  /// Optional GP id → (r, g, b) for map ownership colours; stored on Game and in result.
+  final Map<String, (int r, int g, int b)>? greatPowerColorOverride;
+}
